requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
asyncer==0.0.8
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.5.42
duckduckgo_search==7.2.0
executing==2.1.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jsonschema_rs==0.25.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.28
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langgraph==0.2.60
langgraph-api==0.0.15
langgraph-checkpoint==2.0.8
langgraph-cli==0.1.65
langgraph-sdk==0.1.48
langsmith==0.1.147
litellm==1.51.0
lxml==5.3.0
magicattr==0.1.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.20.1
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.56.1
optuna==4.1.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
primp==0.9.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.4
s3transfer==0.10.4
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.3
structlog==24.4.0
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
